+++ framework/helpers/BaseHtml.php
@@ -1961,7 +1961,7 @@ class BaseHtml
                             if ($v) {
                                 $html .= " $name-$n";
                             }
-                        } else {
+                        } elseif ($v !== null) {
                             $html .= " $name-$n=\"" . static::encode($v) . '"';
                         }
                     }